CH4: L3: Unfollow

+++ sql/queries/feed_follows.sql
@@ -25,8 +25,4 @@ INNER JOIN users ON feed_follows.user_id = users.id
 WHERE feed_follows.user_id = $1;
 
 -- name: DeleteFeedFollow :exec
-DELETE FROM feed_follows
-USING feeds
-WHERE feed_follows.feed_id = feeds.id
-  AND feed_follows.user_id = $1
-  AND feeds.url = $2;
+DELETE FROM feed_follows WHERE feed_id = $1 AND user_id = $2;
